requirements.txt
streamlit
python-dotenv
langchain
langchain-community
langchain-mistralai
mistralai
pypdf
faiss-cpu
